fixed up test structure

+++ server/test/services/zoho/zoho.spec.js
@@ -11,18 +11,20 @@ const expect = chai.expect;
 const should = chai.should();
 chai.use(chaiHttp);
 
-// if (settings.ENVIRONMENT === 'test-all') {
-    describe('zoho tokens', function () {
-        this.timeout(7000);
+describe('zoho tokens', function () {
+    this.timeout(7000);
 
-        afterEach(async function () {
+    afterEach(async function () {
+        if (settings.ENVIRONMENT === 'test-all') {
             console.log('dropping database');
             await mongo.drop();
-        })
+        }
+    })
 
-        describe('using Zoho CRM API', function () {
+    describe('using Zoho CRM API', function () {
 
-            it('should get a new auth token and call the API', async function () {
+        it('should get a new auth token and call the API', async function () {
+            if (settings.ENVIRONMENT === 'test-all') {
                 await zoho.initialize();
                 await zoho.generateAuthTokenfromRefreshToken();
 
@@ -34,12 +36,14 @@ chai.use(chaiHttp);
                 });
 
                 assert(contacts.data.length === 1, "No contacts founds");
-            })
+            }
+        })
 
-            it('should get new auth token', async function () {
+        it('should get new auth token', async function () {
+            if (settings.ENVIRONMENT === 'test-all') {
                 await zoho.initialize();
                 // await zoho.generateAuthTokens();
-            })
+            }
         })
-    });
-// }
+    })
+});
